requirements.txt
openai>=1.0.0
requests>=2.25.1
python-dotenv>=1.0.0
streamlit>=1.30.0
